requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
datasets>=2.14.0
accelerate>=0.24.0
flash-attn>=2.3.0 